Refactoring of HgCommand (and therefore also most HgXXXClients).
HgCommand now extends AbstractShellCommand, which allows to call other executables than hg. This was necessary to prevent code doubling when calling gpg. Essentially, the only difference is that most HgCommand methods have been pulled up to AbstractShellCommand and a new abstract method was introduced, named getExecutable() which must be implemented by each subclass of AbstractShellCommand.

Note: SignWizard now has an icon in team menu, the GPG Executable can be set in the Preferences).