Refactored ChangeSet.setParents(String[]) to filter null parents.
The constructors of Changeset now call setParent(String[]) to set the parents.

Reason:
Hg reports -1:0000000000000000000000000000000000000000 as parent if there is no parent. This is filtered now, as it only clutters the changesets and takes memory we don't need to take. Moreover the filtering prevents that each visualization solution need to implement this filter itself.