Disabling timeout for push and making it configurable in wizard.
Default is "no timeout". If checkbox is enabled, timeout is set to 5 min.

Reason: I had an aborted push due to a timeout when pushing a completely new mercurialeclipse to a repo and I definitely didn't want a timeout then.

ToDo: We should make the defaults adjustable at the MercurialPreferencePage.