Some fixed that should improve performance for big projects.
What would happen if you had, say, 16000 changesets and a huge amount of files (have a look at open solaris or mozlla...)? Getting the whole history won't work. So the plugin now gets only the latest xxx revisions (configurable) of the project and if the file wasn't changed in the last 200 changesets for single files. The result is that there are more calls to hg but the performance shouldn't degrade to badly. The calls for each file only happen, when the label decoration is triggered, so it is lazy loaded.