Added support for GIT-like local branches via the bookmarks extension.
The integration checks if hg help bookmarks yields a result when enabling it on the command line via \"--config extensions.hgext.bookmarks=\". This should succeed if the Mercurial version is newer than 24.10.2008. If the extension is not available, the user is notified.

We now have a \"Manage bookmarks...\" menu item that calls a dialog that enables creation, rename and deletion of bookmarks. The dialog's handler shows an error message if bookmarks is not available.

Moreover, the revision chooser dialog now supports bookmarks. A new tab item has been included that shows all available bookmarks. Selecting a bookmark there and pressing okay will call hg update -C with the bookmark revision. This extra tab item is not created/shown, if the bookmarks extension is not available.