Fix thread-safety and other issues in HgConsole and HgConsoleFactory.
Make HgConsoleFactory stateless since an instance is created by Eclipse wheneve the open console action is invoked. Move most of the code to HgConsoleHolder.

Remove static fields and methods from HgConsoleHolder since it's a singleton. Simplify code by removing a few fields that can be replaced with methods. Document that the class is not thread-safe and has to be called from the UI thread.

Change DefaultConfiguration to call HgConsole methods from the UI thread. Instead of storing the console field, retrieve it from HgConsoleHolder when required. This means that the code will work correctly if a HgConsole is closed and then a new one is created.

Adapt ConsoleRemoveAction to the fact that HgConsoleHolder.closeConsole is now an instance method.