Don't activate forest and subversion extensions automatically.
They must be activated in .hgrc. Only extensions distributed together with Mercurial are auto-activated now as we can't know the correct path for any other extensions.

Later it might be the better option to read hgrc and offer a way to edit extension paths via a dialog. This shouldn't alter the hgrc, though.