Repo view and push/pull pages: changed overall behavior to always remember last user/pass for a given URL. This behavior occurs from any page that allows users to enter in a URL.
All repositories are automatically remembered (and added to repository database)
automatically. They may be removed from the repository view later, of course.
The advantage of this is:
    1. repository user/pass are always remembered regardless of location
    2. all these things are viewable from repository view

Note that the previous behavior was to only remember repository locations
that were entered in the "clone repo create project wizard" and a few other
places.

Changed HgRepositoryLocation to be more immutable.