Make encoding settings configurable
Filenames, content and metadata are not always transcoded correctly.
In order to fix this, we now determine the encoding to use by following this
algorithm:

1) Check Eclipse project encoding.
2) Check for HGENCODING environment variable to obtain encoding.
3) Check for JVM default charset

All this stuff is managed using the HgRoot class. The Eclipse specific stuff resides in
MercurialTeamProvider in order to avoid cyclic dependencies. Manually setting encodings
on a HgRoot object always override the automatically determined ones.

The search4HgRoot methods in MercurialUtilities have been superseded by the ones in
MercurialTeamProvider, resulting in only one location to remember for asking for a
configured HgRoot.

HgRoot also defines a fallback encoding if it is set in the repositories .hgrc configuration file.
Else it is set to the windows locale cp1251 - don't sue me for this...
The global preferences aren't read so far.