minor fix for logbatchsize in changeset table + note how to improve performance.
the switch to dialog is slow when the cache already is filled with a lot of revisions. on large repositories this is a major UI bottleneck, as the table contains all revisions of the repository. A filter should be applied that only displays the so far requested revisions.

It might be necessary to supply a jump-function as well.