Merge with62844f50ed6a14ada6b47fb150590e09b4942fac This merge with the sync-redesign and pref_optimization branches is the main content for the next 1.5.1 (or 1.6.0) version.
The biggest and most important change is related to the way how eclipse plugin run hg operations. Instead to threat projects as the center of hg universe, we run most of the hg operations on hg root.

This brings an immense performance benefit for these use cases where multiple projects are contained in one hg root. In the case of a repository with 50 projects and 50000 files it means that the status/remote sync commands run in seconds instead of minutes.
Fixes for: issue #10646, issue #10811, issue #10810, issue #10803, issue #10682, issue #10679, issue #10484