changed commit names not to use repository usernames if there are configured mercurial usernames.
usernames are not full names, mostly don't contain email addresses and in general are only a poor substitute for a full name. if a person has configured his mercurial this configuration seems to be always better than a repository username - if someone wants to set a different name for a repo he still has two options: setting it in the property page or in the repo hgrc in .hg.