Introduce preference for choosing how the default commit username should be determined (rel: #11007)
If checked (the default), the default username is taken from the hg config, or, if not set the username set in the environent is used. If not checked, the default username is the repository username used for authentication. This only happens once for each repository.

After the first time providing a username to Mercurial, this username is stored as a property associated with the repository and used and updated afterwards so that the default determination doesn't apply anymore.